requirements.txt
openpyxl
langchain 
openai 
pandas
groq 
langchain-community 
pandas 
python-dotenv
gspread 
pandas 
oauth2client